remove SplitPosition. DataIterator and IcebergSourceSplit tracks and expose fileOffset and recordOffset directly.
