--- flink/v1.13/flink/src/main/java/org/apache/iceberg/flink/source/SplitPosition.java
@@ -1,95 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iceberg.flink.source;
-
-import java.io.Serializable;
-import java.util.Objects;
-import org.apache.flink.annotation.Internal;
-import org.apache.iceberg.CombinedScanTask;
-import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
-
-/**
- * A mutable class that defines the read position
- * <ul>
- *   <li>file offset in the list of files in a {@link CombinedScanTask}</li>
- *   <li>record offset within a file</li>
- * </ul>
- */
-@Internal
-public class SplitPosition implements Serializable {
-
-  private static final long serialVersionUID = 1L;
-
-  private int fileOffset;
-  private long recordOffset;
-
-  public SplitPosition(int fileOffset, long recordOffset) {
-    this.fileOffset = fileOffset;
-    this.recordOffset = recordOffset;
-  }
-
-  void advanceFile() {
-    this.fileOffset += 1;
-    this.recordOffset = 0L;
-  }
-
-  void advanceRecord() {
-    this.recordOffset += 1L;
-  }
-
-  public void set(int newFileOffset, long newRecordOffset) {
-    this.fileOffset = newFileOffset;
-    this.recordOffset = newRecordOffset;
-  }
-
-  public int fileOffset() {
-    return fileOffset;
-  }
-
-  public long recordOffset() {
-    return recordOffset;
-  }
-
-  @Override
-  public boolean equals(Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    final SplitPosition that = (SplitPosition) o;
-    return fileOffset == that.fileOffset &&
-        recordOffset == that.recordOffset;
-  }
-
-  @Override
-  public int hashCode() {
-    return Objects.hash(fileOffset, recordOffset);
-  }
-
-  @Override
-  public String toString() {
-    return MoreObjects.toStringHelper(this)
-        .add("fileOffset", fileOffset)
-        .add("recordOffset", recordOffset)
-        .toString();
-  }
-}
